requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
flask-restx==1.3.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
azure-ai-ml==1.16.1
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==31.0.0
azure-mgmt-core==1.4.0
azure-mgmt-network==25.4.0
azure-mgmt-resource==23.1.1
azure-mgmt-search==9.1.0
azure-mgmt-storage==21.1.0
azure-monitor-opentelemetry-exporter==1.0.0b25
azure-storage-blob==12.20.0
azure-storage-file-datalake==12.15.0
azure-storage-file-share==12.16.0
openai==1.31.1
requests==2.32.2
requests-oauthlib==2.0.0
aiofiles==23.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytest==7.4.0